server_python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
websockets==12.0
psutil==5.9.8
pytest>=8.2.0,<9.0.0
pytest-asyncio==0.24.0

# Scientific computing (used by processing scripts)
scipy>=1.9.0

# XGBoost leeway model (gp50 2_processing calibration)
xgboost>=2.0.0
joblib>=1.3.0

# Core dependencies for api_utils (used by GP50 normalization)
pandas>=1.5.0
numpy>=1.20.0
requests>=2.28.0
python-dateutil>=2.8.0

# InfluxDB client (used by api_utils for GP50 normalization)
influxdb-client>=1.36.0